docs/template_config_files/scripts/requirements.txt
requests
GitPython
pytest
pytest-reportportal
reportportal-client
